move InferencePhase into model_inference.cpp; delete header
InferencePhase is a template used only by run_inference. Moving it into
the anonymous namespace of model_inference.cpp keeps all llama.cpp calls
inside model*.cpp per the core-layer rule, and removes the only header
in src/core/ that included llama.h.
